the i variable needed to be reworked
It was sending both signals, because break just breaks out of the while loop, and not the parent if statement. So it was still triggering the if statement below the while, and when setting i to 0 during the if i >= 20 part, it would cause the if i < 20 to trigger as well.