reworked code for function
Made a function to do everything the old if, while, if, if statements. Now if the button is pressed, it calls the function to begin counting i for long press, and if shorter than 2 seconds sends the value that was passed to it, reading the signalWire that was also passed to it. Need to test.