added a 3 sec delay for suspend/resume
Figured it would be best to have a delay, since when turning the car initially on, you have to pass "acc" in the cars housing, triggering startup. Then it cuts for ignition, and back on again once the car is started. So 3 seconds should be good enough to cover that gap without being too long.